+++ py/k8s_util.py
@@ -9,50 +9,6 @@ import time
 from kubeflow.testing import util
 from kubernetes import client as k8s_client
 from kubernetes.client import rest
-from py import tf_job_client
-
-
-def wait_for_delete(client,
-                    namespace,
-                    name,
-                    version="v1alpha1",
-                    timeout=datetime.timedelta(minutes=5),
-                    polling_interval=datetime.timedelta(seconds=30),
-                    status_callback=None):
-  """Wait for the specified job to be deleted.
-
-  Args:
-    client: K8s api client.
-    namespace: namespace for the job.
-    name: Name of the job.
-    timeout: How long to wait for the job.
-    polling_interval: How often to poll for the status of the job.
-    status_callback: (Optional): Callable. If supplied this callable is
-      invoked after we poll the job. Callable takes a single argument which
-      is the job.
-  """
-  crd_api = k8s_client.CustomObjectsApi(client)
-  end_time = datetime.datetime.now() + timeout
-  while True:
-    try:
-      results = crd_api.get_namespaced_custom_object(
-        tf_job_client.TF_JOB_GROUP, version, namespace,
-        tf_job_client.TF_JOB_PLURAL, name)
-    except rest.ApiException as e:
-      if e.status == httplib.NOT_FOUND:
-        return
-      logging.exception("rest.ApiException thrown")
-      raise
-    if status_callback:
-      status_callback(results)
-
-    if datetime.datetime.now() + polling_interval > end_time:
-      raise util.TimeoutError(
-        "Timeout waiting for job {0} in namespace {1} to be deleted.".format(
-          name, namespace))
-
-    time.sleep(polling_interval.seconds)
-
 
 def log_pods(pods):
   """Log information about pods."""
@@ -228,5 +184,3 @@ def parse_events(events):
       services.add(name)
 
   return pods, services
-
-
